+++ src/query-builder/UpdateQueryBuilder.ts
@@ -62,6 +62,33 @@ export class UpdateQueryBuilder<Entity> extends QueryBuilder<Entity> {
         return this;
     }
 
+    /**
+     * Adds new AND WHERE with conditions for the given ids.
+     */
+    whereInIds(ids: any[]): this {
+        const [whereExpression, parameters] = this.createWhereIdsExpression(ids);
+        this.andWhere(whereExpression, parameters);
+        return this;
+    }
+
+    /**
+     * Adds new AND WHERE with conditions for the given ids.
+     */
+    andWhereInIds(ids: any[]): this {
+        const [whereExpression, parameters] = this.createWhereIdsExpression(ids);
+        this.andWhere(whereExpression, parameters);
+        return this;
+    }
+
+    /**
+     * Adds new OR WHERE with conditions for the given ids.
+     */
+    orWhereInIds(ids: any[]): this {
+        const [whereExpression, parameters] = this.createWhereIdsExpression(ids);
+        this.orWhere(whereExpression, parameters);
+        return this;
+    }
+
     // -------------------------------------------------------------------------
     // Protected Methods
     // -------------------------------------------------------------------------
@@ -72,6 +99,7 @@ export class UpdateQueryBuilder<Entity> extends QueryBuilder<Entity> {
     protected createUpdateExpression() {
         const valuesSet = this.getValueSets();
 
+        // prepare columns and values to be updated
         const updateColumnAndValues: string[] = [];
         Object.keys(valuesSet).forEach(columnProperty => {
             const column = this.expressionMap.mainAlias!.metadata.findColumnWithPropertyName(columnProperty);